requirements.txt
Flask==2.3.3
flask-cors==4.0.0
pywebview==4.2.2
